requirements.txt
bcrypt
fastapi
pydantic
pyjwt